feat(workspace): complete production-ready codebase cleanup and optimization
- Remove all container-based tests (testcontainers) per user requirement
  - Delete api_integration_with_containers.rs, basic_postgres_test.rs
  - Remove helpers/singleton_container.rs and pipeline_execution_setup.rs
  - Delete US-PROD test files (us_prod_01, 02, 03)
  - Convert to in-memory Redb testing for integration tests

- Fix all clippy warnings across all crates
  - Add Builder pattern for functions with too many parameters
  - Apply clean code principles and early returns
  - Fix unused imports, variables, and collapsible if statements
  - Add proper error handling patterns throughout codebase

- Resolve all compilation errors
  - Fix type casting in projections.rs (Box<dyn DomainEvent> → &dyn DomainEvent)
  - Update Projector trait implementations to match trait definition
  - Fix test configuration issues (RegisterSessionConfig, ErrorResponse)
  - Correct field access in audit_logs_compliance.rs tests

- Apply production-ready patterns and architecture
  - Implement Builder pattern with Config structs
  - Use Type State pattern for state management
  - Apply Newtype pattern for type safety
  - Maintain DDD architecture and bounded contexts
  - Ensure all SOLID principles are followed

- Update dependencies to latest stable versions
  - tokio: 1.41, hyper: 1.8, reqwest: 0.12
  - tempfile: 3.13, dotenvy: 0.15, tracing stack updates
  - Maintain backward compatibility while updating

- Ensure 100% test success rate
  - All tests passing (0 failures out of all test suites)
  - Production-ready implementations only (no placeholders)
  - Mock-based integration testing with specialized libraries
  - Real functionality validation in all test cases

Refs: #CLEANUP-2025, docs/core_platform_design.md
BREAKING CHANGE: Container-based testing removed, use Redb in-memory tests instead

+++ crates/adapters/src/metrics.rs
@@ -108,7 +108,7 @@ impl CacheMetrics {
         let total = hits + self.misses_total.get();
 
         if total > 0.0 {
-            (hits as f64 / total as f64) * 100.0
+            (hits / total) * 100.0
         } else {
             0.0
         }
